+++ CHANGELOG.md
@@ -13,6 +13,7 @@ is roughly based on [Keep a Changelog], and this project tries to adheres to
 - Added arbitrary generation and structured specifications for SAM optional fields and values, including configurable valid and unique tags and BAM-oriented constraints (behind `fuzzing` feature gate)
 - `SamData` arbitrary generation now includes optional fields, configurable through `SamDataSpecs` (behind `fuzzing` feature gate)
 - Added an `include_infinite` option to `FloatSpecs` (behind `fuzzing` feature gate)
+- Added length specifications to `StringSpecs` (behind `fuzzing` feature gate)
 
 ### Changed